Fix treatment of negative base in `__pow__` backward pass
We weren't handling calculation of the gradient of x^y correctly. The
derivative is x^y * log(x), which because of log(x) is only defined
when x > 0. We were always attempting to calculate log(x), resulting in
a math domain error when using this in the MSELoss calculation. We fix
this by adding a conditional to only compute the exponent's gradient
when the x is positive.